Split out main into crate mods
