+++ src/init/mod.rs
@@ -1,0 +1,8 @@
+pub mod agent;
+pub mod args;
+pub mod activation;
+mod otlp_exporter;
+mod datadog_exporter;
+
+#[cfg(feature = "pprof")]
+pub mod pprof;
