+++ src/lib.rs
@@ -7,3 +7,4 @@ pub mod receivers;
 pub mod semconv;
 pub mod telemetry;
 pub mod topology;
+pub mod init;
